requirements.txt
fastapi==0.110.1
uvicorn[standard]==0.29.0
sentence-transformers==2.6.1
httpx==0.27.0
openai==1.30.1
python-dotenv